requirements.txt
traceloop-sdk==0.25.1
fastapi==0.111.0
openai==1.35.13
instructor==1.3.4
pydantic==2.8.2
tantivy==0.22.0
braintrust==0.0.142
langfuse==2.39.0
datasets==2.20.0
lancedb==0.9.0
scikit-learn==1.5.1